+++ lorawan-device/src/async_device/lora_radio.rs
@@ -3,7 +3,7 @@ use super::Timings;
 
 use lora_phy::mod_params::{BoardType, RadioError};
 use lora_phy::mod_traits::RadioKind;
-use lora_phy::{DelayUs, LoRa};
+use lora_phy::{DelayNs, LoRa};
 
 /// LoRa radio using the physical layer API in the external lora-phy crate.
 ///
@@ -12,7 +12,7 @@ use lora_phy::{DelayUs, LoRa};
 pub struct LoRaRadio<RK, DLY, const P: u8, const G: i8 = 0>
 where
     RK: RadioKind,
-    DLY: DelayUs,
+    DLY: DelayNs,
 {
     pub(crate) lora: LoRa<RK, DLY>,
     rx_pkt_params: Option<lora_phy::mod_params::PacketParams>,
@@ -20,7 +20,7 @@ where
 impl<RK, DLY, const P: u8, const G: i8> LoRaRadio<RK, DLY, P, G>
 where
     RK: RadioKind,
-    DLY: DelayUs,
+    DLY: DelayNs,
 {
     pub fn new(lora: LoRa<RK, DLY>) -> Self {
         Self { lora, rx_pkt_params: None }
@@ -31,7 +31,7 @@ where
 impl<RK, DLY, const P: u8, const G: i8> Timings for LoRaRadio<RK, DLY, P, G>
 where
     RK: RadioKind,
-    DLY: DelayUs,
+    DLY: DelayNs,
 {
     fn get_rx_window_offset_ms(&self) -> i32 {
         match self.lora.get_board_type() {
@@ -68,7 +68,7 @@ impl From<RadioError> for Error {
 impl<RK, DLY, const P: u8, const G: i8> PhyRxTx for LoRaRadio<RK, DLY, P, G>
 where
     RK: RadioKind,
-    DLY: DelayUs,
+    DLY: DelayNs,
 {
     type PhyError = Error;